requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
jinja2
google-auth
gunicorn
sqlalchemy
PyJWT
requests
yfinance
pytz
apscheduler
python-dotenv>=1.0
kiteconnect
fastapi_utils
typing_inspect
openpyxl
pandas
pydantic
resend>=2.0.0
qrcode[pil]
stripe
ta
email-validator
openpyxl==3.1.5
fastapi==0.110.0
uvicorn==0.27.1
